requirements.txt
torch==1.5.1+cu101
torchvision==0.6.1+cu101
Pillow~=7.2.0